update prod config for deploy
